#6 rearranged code
