+++ src/lib.rs
@@ -16,13 +16,8 @@ extern crate uuid;
 
 extern crate url;
 
-mod model;
 mod auth;
 mod nakadi;
-
-/// A `SubscriptionId` is used to guaratee a continous flow of events for a client.
-#[derive(Clone, Debug)]
-pub struct SubscriptionId(pub String);
 
 /// Describes what to do after a batch has been processed.
 ///
@@ -30,10 +25,4 @@ pub struct SubscriptionId(pub String);
 pub enum AfterBatchAction {
     Continue,
     Abort { reason: String },
-}
-
-pub struct EventType<'a>(&'a str);
-
-pub trait Handler {
-    fn handle_batch(&self, event_type: EventType, data: &[u8]) -> AfterBatchAction;
 }
