--- src/model.rs
@@ -1,1 +1,0 @@
-
