+++ src/nakadi/subscription/mod.rs
@@ -1,0 +1,5 @@
+
+
+pub mod stream;
+pub mod processor;
+pub mod model;
